fix(api): resolve lint findings in subagent tool projection and tombstone guard test
Break the import cycle between spawn-subagents-tool.ts and
subagent-tools.ts (introduced by the buffered-approval change) by
extracting the shared symbols both sides need — the tool name constant
and the proposal-buffer types/factory — into a new
subagent-tool-shared.ts module. Both files now import from the shared
module instead of from each other, and each keeps a local re-export of
its previously-exported names so downstream call sites are unaffected.

Also drop the redundant `async` on the proposal wrapper's `execute`
(it never awaited anything) and normalize `"utf8"` to the canonical
`"utf-8"` encoding identifier in the delete-version tombstone-guard
test.
